Refactor: Extract linebased adapters to separate module
Reorganizes adapter architecture to clearly separate concerns:

**New structure:**
- adapters.rs: Generic TokenStream ↔ Vec<Token> adapters (flat conversion)
- adapters_linebased.rs: LineBased-specific TokenStream ↔ LineContainer adapters

**Changes:**
- Created src/lex/pipeline/adapters_linebased.rs with:
  - line_container_to_token_stream()
  - token_stream_to_line_tokens()
  - token_stream_to_line_container()
  - adapt_linebased_parser()
  - All related tests (467 lines)

- Cleaned adapters.rs:
  - Removed linebased-specific code
  - Kept generic flat adapters and reference parser adapter

- Updated imports:
  - linebased/pipeline.rs now imports from adapters_linebased
  - Added module declaration to pipeline/mod.rs

**Architecture confirmed:**
- Transformation pipeline is 100% adapter-free (TokenStream throughout)
- Adapters only exist at architectural boundaries:
  - Generic: Vec<Token> ↔ TokenStream
  - LineBased: TokenStream ↔ LineContainer (lexer→parser boundary)

✅ All 381 tests passing
✅ Zero clippy warnings
